+++ R/status.R
@@ -261,7 +261,35 @@ renv_status_check_consistent <- function(lockfile, library, used) {
 
 }
 
+renv_status_check_enabled <- function(parent) {
+
+  # get the name of the calling function
+  ok <- is.call(parent) && length(parent) && is.symbol(parent[[1L]])
+  if (!ok)
+    return(TRUE)
+
+  invoker <- as.character(parent[[1L]])
+  parts <- strsplit(invoker, "_", fixed = TRUE)[[1L]]
+  if (length(parts) < 3L)
+    return(TRUE)
+
+  # check the relevant config option
+  scope <- parts[[2L]]
+  name <- paste(tail(parts, n = -2L), collapse = "_")
+  value <- renv_config_get(
+    name    = name,
+    scope   = scope,
+    default = TRUE
+  )
+
+  truthy(value, default = TRUE)
+
+}
+
 renv_status_check_initialized <- function(project, library = NULL, lockfile = NULL) {
+
+  if (!renv_status_check_enabled(sys.call()))
+    return(TRUE)
 
   # only done if library and lockfile are NULL; that is, if the user
   # is calling `renv::status()` without arguments
@@ -304,6 +332,9 @@ renv_status_check_initialized <- function(project, library = NULL, lockfile = NU
 
 renv_status_check_synchronized <- function(lockfile, library) {
 
+  if (!renv_status_check_enabled(sys.call()))
+    return(TRUE)
+
   lockfile <- renv_lockfile_records(lockfile)
   library <- renv_lockfile_records(library)
 
@@ -329,6 +360,9 @@ renv_status_check_synchronized <- function(lockfile, library) {
 }
 
 renv_status_check_version <- function(lockfile) {
+
+  if (!renv_status_check_enabled(sys.call()))
+    return(TRUE)
 
   version <- lockfile$R$Version
   if (renv_version_eq(version, getRversion(), n = 2L))